+++ R/dummy.R
@@ -118,7 +118,7 @@ step_dummy <-
            id = rand_id("dummy")) {
 
     if (lifecycle::is_present(preserve)) {
-      lifecycle::deprecate_soft(
+      lifecycle::deprecate_warn(
         "0.1.16",
         "step_dummy(preserve = )",
         "step_dummy(keep_original_cols = )"