+++ R/kpca.R
@@ -1,69 +1,30 @@
 #' Kernel PCA Signal Extraction
 #'
-#' `step_kpca` a *specification* of a recipe step that
+#' `step_kpca` creates a *specification* of a recipe step that
 #'  will convert numeric data into one or more principal components
 #'  using a kernel basis expansion.
 #'
 #' @inheritParams step_pca
 #' @inheritParams step_center
-#' @param num_comp The number of PCA components to retain as new
-#'  predictors. If `num_comp` is greater than the number of columns
-#'  or the number of possible components, a smaller value will be
-#'  used.
-#' @param options A list of options to
-#'  [kernlab::kpca()]. Defaults are set for the arguments
-#'  `kernel` and `kpar` but others can be passed in.
-#'  **Note** that the arguments `x` and `features`
-#'  should not be passed here (or at all).
-#' @param res An S4 [kernlab::kpca()] object is stored
-#'  here once this preprocessing step has be trained by
-#'  [prep.recipe()].
+#' @param options A list of options to [kernlab::kpca()]. Defaults are set for
+#'  the arguments `kernel` and `kpar` but others can be passed in.
+#'  **Note** that the arguments `x` and `features` should not be passed here
+#'  (or at all).
+#' @param res An S4 [kernlab::kpca()] object is stored here once this
+#'  preprocessing step has be trained by [`prep()`][prep.recipe()].
 #' @template step-return
-#' @keywords internal
+#' @family {multivariate transformation steps}
 #' @export
-#' @details Kernel principal component analysis (kPCA) is an
-#'  extension of a PCA analysis that conducts the calculations in a
-#'  broader dimensionality defined by a kernel function. For
-#'  example, if a quadratic kernel function were used, each variable
-#'  would be represented by its original values as well as its
-#'  square. This nonlinear mapping is used during the PCA analysis
-#'  and can potentially help find better representations of the
-#'  original data.
+#' @details
+#' When performing kPCA with `step_kpca()`, you must choose the kernel
+#' function (and any important kernel parameters). This step uses the
+#' \pkg{kernlab} package; the reference below discusses the types of kernels
+#' available and their parameter(s). These specifications can be made in the
+#' `kernel` and `kpar` slots of the `options` argument to `step_kpca()`.
+#' Consider using [step_kpca_rbf()] for a radial basis function kernel or
+#' [step_kpca_poly()] for a polynomial kernel.
 #'
-#' This step requires the \pkg{dimRed} and \pkg{kernlab} packages.
-#' If not installed, the step will stop with a note about installing
-#' these packages.
-#'
-#' As with ordinary PCA, it is important to standardize the
-#'  variables prior to running PCA (`step_center` and
-#'  `step_scale` can be used for this purpose).
-#'
-#' When performing kPCA, the kernel function (and any important
-#'  kernel parameters) must be chosen. The \pkg{kernlab} package is
-#'  used and the reference below discusses the types of kernels
-#'  available and their parameter(s). These specifications can be
-#'  made in the `kernel` and `kpar` slots of the
-#'  `options` argument to `step_kpca`.
-#'
-#' The argument `num_comp` controls the number of components that
-#'  will be retained (the original variables that are used to derive
-#'  the components are removed from the data). The new components
-#'  will have names that begin with `prefix` and a sequence of
-#'  numbers. The variable names are padded with zeros. For example,
-#'  if `num_comp < 10`, their names will be `kPC1` -
-#'  `kPC9`. If `num_comp = 101`, the names would be
-#'  `kPC001` - `kPC101`.
-#'
-#' When you [`tidy()`] this step, a tibble with column `terms` (the
-#'  selectors or variables selected) is returned.
-#'
-#' @references Scholkopf, B., Smola, A., and Muller, K. (1997).
-#'  Kernel principal component analysis. *Lecture Notes in
-#'  Computer Science*, 1327, 583-588.
-#'
-#' Karatzoglou, K., Smola, A., Hornik, K., and Zeileis, A. (2004).
-#'  kernlab - An S4 package for kernel methods in R. *Journal
-#'  of Statistical Software*, 11(1), 1-20.
+#' @template kpca-info
 #'
 #' @examples
 #' library(modeldata)
@@ -92,10 +53,6 @@
 #'   tidy(kpca_trans, number = 3)
 #'   tidy(kpca_estimates, number = 3)
 #' }
-#' @seealso [step_pca()] [step_ica()]
-#'   [step_isomap()] [recipe()] [prep.recipe()]
-#'   [bake.recipe()]
-#'
 step_kpca <-
   function(recipe,
            ...,
@@ -106,16 +63,11 @@ step_kpca <-
            options = list(kernel = "rbfdot",
                           kpar = list(sigma = 0.2)),
            prefix = "kPC",
+           keep_original_cols = FALSE,
            skip = FALSE,
            id = rand_id("kpca")) {
 
     recipes_pkg_check(required_pkgs.step_kpca())
-    rlang::inform(
-      paste(
-        "`step_kpca()` is deprecated in favor of either `step_kpca_rbf()`",
-        "or `step_kpca_poly()`. It will be removed in future versions."
-      )
-    )
 
     add_step(
       recipe,
@@ -127,6 +79,7 @@ step_kpca <-
         res = res,
         options = options,
         prefix = prefix,
+        keep_original_cols = keep_original_cols,
         skip = skip,
         id = id
       )
@@ -134,7 +87,8 @@ step_kpca <-
 }
 
 step_kpca_new <-
-  function(terms, role, trained, num_comp, res, options, prefix, skip, id) {
+  function(terms, role, trained, num_comp, res, options, prefix,
+           keep_original_cols, skip, id) {
     step(
       subclass = "kpca",
       terms = terms,
@@ -144,13 +98,13 @@ step_kpca_new <-
       res = res,
       options = options,
       prefix = prefix,
+      keep_original_cols = keep_original_cols,
       skip = skip,
       id = id
     )
   }
 
 #' @export
-#' @keywords internal
 prep.step_kpca <- function(x, training, info = NULL, ...) {
   col_names <- recipes_eval_select(x$terms, training, info)
   check_type(training[, col_names])
@@ -182,13 +136,13 @@ prep.step_kpca <- function(x, training, info = NULL, ...) {
     options = x$options,
     res = kprc,
     prefix = x$prefix,
+    keep_original_cols = get_keep_original_cols(x),
     skip = x$skip,
     id = x$id
   )
 }
 
 #' @export
-#' @keywords internal
 bake.step_kpca <- function(object, new_data, ...) {
   if (object$num_comp > 0) {
     pca_vars <- colnames(environment(object$res@apply)$indata)
@@ -198,7 +152,11 @@ bake.step_kpca <- function(object, new_data, ...) {
     comps <- comps[, 1:object$num_comp, drop = FALSE]
     comps <- check_name(comps, new_data, object)
     new_data <- bind_cols(new_data, as_tibble(comps))
-    new_data <- new_data[, !(colnames(new_data) %in% pca_vars), drop = FALSE]
+    keep_original_cols <- get_keep_original_cols(object)
+
+    if (!keep_original_cols) {
+      new_data <- new_data[, !(colnames(new_data) %in% pca_vars), drop = FALSE]
+    }
   }
   as_tibble(new_data)
 }
@@ -222,7 +180,6 @@ print.step_kpca <- function(x, width = max(20, options()$width - 40), ...) {
 
 #' @rdname tidy.recipe
 #' @export
-#' @keywords internal
 tidy.step_kpca <- function(x, ...) {
   if (is_trained(x)) {
     if (x$num_comp > 0) {