+++ NEWS.md
@@ -27,6 +27,15 @@
 
 * A bug was fixed where imputed values via bagged trees would have the wrong levels.
 
+* `step_kpca()` was un-deprecated and gained the `keep_original_cols` argument.
+
+* The deprecation of the `preserve` argument to `step_pls()` and `step_dummy()` was escalated from a soft deprecation to regular deprecation. 
+
+* The deprecation of the `options` argument to `step_nzv()` was escalated to a deprecation error.
+
+* The `tidy()` methods for objects from `check_new_values()`, `check_class()` and `step_nnmf()` are now exported.
+
+
 # recipes 0.1.16
 
 ## New Steps